+++ CHANGELOG.md
@@ -30,8 +30,13 @@ WorldEditArt::requireVersion($this->getServer(), 4, 1, 0);
     - Except:
       - `//cyl normalize false` (`false` for `preserveLength`) is not properly implemented, and temporarily redirects to `//cyl normalize true` due to technical difficulties that can be overcome _had I tried harder learning vector mathematics_.
     - **UNDOCUMENTED: Commands**
+  - Ellipsoid selection (commands and wands)
   - Manipulation commands
     - //set
+      - Only command prsing implemented
+      - **UNDOCUMENTED**
+    - //replace
+      - Only command prsing implemented
       - **UNDOCUMENTED**
 
 [Unreleased]: https://github.com/LegendOfMCPE/WorldEditArt/compare/delta/v3.0...HEAD